// csm.web.app/src/app/shared/misc/enums/modal-id.enum.ts
/*
 * *************************************************************************
 *
 * Copyright: Robert Bosch Power Tools GmbH, 2023
 *
 * *************************************************************************
 */

export enum ModalIdEnum {
    CalendarExport = 'calendar-export',
    ConfirmationDialog = 'confirm-dialog',
    ConfirmDayCardStatusChangeWithReasons = 'confirm-day-card-status-change-with-reasons',
    ConfirmMultipleDayCardStatusChangeWithReasons = 'confirm-multiple-day-card-status-change-with-reasons',
    CreateDayCard = 'create-day-card',
    CreatePAT = 'create-pat',
    CreateTask = 'create-task',
    CreateTopic = 'create-topic',
    LegalDocuments = 'legal-documents',
    PrivacySettings = 'privacy-settings',
    ProjectCopy = 'project-copy',
    ProjectImport = 'project-import',
    ProjectExport = 'project-export',
    UpdateDayCard = 'update-day-card',
    UpdateMilestone = 'update-milestone',
    UpdatePAT = 'update-pat',
    UpdateTask = 'update-task',
    UpdateConstraints = 'constraints',
    UpdateWorkingDaysHoliday = 'update-working-days-holiday',
}