// csm.web.app/src/app/project/project-common/enums/calendar-selection-action.enum.ts
/*
 * *************************************************************************
 *
 * Copyright: Robert Bosch Power Tools GmbH, 2023
 *
 * *************************************************************************
 */

export enum CalendarSelectionActionEnum {
    Move = 'MOVE',
    Copy = 'COPY',
}